Added GET by username and DELTE endpoint. Fixed integration tests

+++ Application/src/main/java/com/kenzie/appserver/service/ReviewService.java
@@ -1,22 +1,23 @@
 package com.kenzie.appserver.service;
 
-import com.amazonaws.services.dynamodbv2.model.ConditionalCheckFailedException;
 import com.kenzie.appserver.config.ReviewCache;
 import com.kenzie.appserver.exceptions.ReviewNotFoundException;
 import com.kenzie.appserver.repositories.ReviewRepository;
 import com.kenzie.appserver.repositories.model.ReviewEntity;
 import com.kenzie.appserver.repositories.model.ReviewPrimaryKey;
 import com.kenzie.appserver.service.model.Review;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ReviewService {
     private final ReviewRepository reviewRepository;
     private final ReviewCache cache;
-
+    @Autowired
     public ReviewService(ReviewRepository reviewRepository, ReviewCache cache) {
         this.reviewRepository = reviewRepository;
         this.cache = cache;
@@ -43,6 +44,13 @@ public class ReviewService {
         return reviews;
     }
 
+    public List<Review> getAllByUsername(String username) {
+        List<ReviewEntity> reviewEntities = reviewRepository.findAllByUsername(username);
+        List<Review> reviews = new ArrayList<>();
+        reviewEntities.forEach(entity -> reviews.add(convertToReview(entity)));
+        return reviews;
+    }
+
     public Review createReview(Review review) {
         review.calculateAndSetTotalRating();
         review.setDatePosted();
@@ -54,25 +62,28 @@ public class ReviewService {
 
 
     public Review updateReview(Review review) {
-        review.calculateAndSetTotalRating();
-        ReviewEntity entity = convertToEntity(review);
-        try{
+        Optional<ReviewEntity> result = reviewRepository.findById(new ReviewPrimaryKey(review.getTeacherName(), review.getDatePosted()));
+        if (!result.isPresent()) {
+            throw new ReviewNotFoundException();
+        } else {
+            review.calculateAndSetTotalRating();
+            ReviewEntity entity = convertToEntity(review);
             reviewRepository.save(entity);
             cache.evictAllReviewsByCourseTitle(entity.getCourseTitle());
-        } catch (ConditionalCheckFailedException e){
-            throw new ReviewNotFoundException();
         }
+
         return review;
     }
 
     public void deleteReview(Review review) {
-        /* TODO check if review exists before deleting. throw custom exception if not found.
-            Reason: DynamoDB does not throw an exception if item does not exist so no indication if successful
-        */
-
-        ReviewEntity entity = convertToEntity(review);
-        reviewRepository.delete(entity);
-        cache.evictAllReviewsByCourseTitle(entity.getCourseTitle());
+        Optional<ReviewEntity> result = reviewRepository.findById(new ReviewPrimaryKey(review.getTeacherName(), review.getDatePosted()));
+        if (!result.isPresent()) {
+            throw new ReviewNotFoundException();
+        } else {
+            ReviewEntity entity = convertToEntity(review);
+            reviewRepository.delete(entity);
+            cache.evictAllReviewsByCourseTitle(entity.getCourseTitle());
+        }
     }
